refactor: rename scss files to to styles.x from index.x
