+++ src/components/Carousel/styles.module.scss
@@ -1,3 +1,5 @@
+@import "../../styles/mixins";
+
 .button {
   background: none;
   border: none;
@@ -30,6 +32,11 @@
   transition: color var(--ifm-transition-fast) var(--ifm-transition-timing-default);
   width: 3rem;
 
+  @include tablet-and-up {
+    height: 3rem;
+    width: 3rem;
+  }
+
   &:hover {
     fill: var(--ifm-color-primary)
   }
@@ -40,7 +47,6 @@
   display: inline-block;
   filter: alpha(opacity=30);
   opacity: 0.3;
-  //box-shadow: 1px 1px 2px rgba(255, 255, 255, 0.9);
   border-radius: 50%;
   width: 8px;
   height: 8px;
