+++ src/components/Main/Main.tsx
@@ -1,7 +1,7 @@
 import React, { FC, PropsWithChildren } from 'react';
 
 // styles
-import styles from './index.module.scss';
+import styles from './styles.module.scss';
 
 const Main: FC<PropsWithChildren> = ({ children }: PropsWithChildren) => (
   <main className={styles.main}>{children}</main>
