+++ src/components/Main/styles.module.scss
