+++ src/components/SectionTitle/styles.module.scss
